move packq to a stand-alone cpp-file to accelerate compile
