+++ intel_extension_for_transformers/llm/operator/csrc/dispatcher/include/jblas_weightonly_dispatcher.hpp
@@ -58,6 +58,14 @@ struct woq_runtime_ctx {
   jblas::storage::gemm::IWeightBase* deseries_wei;
 };
 
+static std::map<std::string, JBLAS_DTYPE> wei2jblasdt_map{
+    {"int4_clip", JBLAS_DTYPE::S4_CLIP}, {"int4_fullrange", JBLAS_DTYPE::S4_FULLRANGE},
+    {"nf4", JBLAS_DTYPE::F4_NF4},        {"fp4_e2m1_bnb", JBLAS_DTYPE::F4_BNB},
+    {"fp4_e2m1", JBLAS_DTYPE::F4_E2M1},  {"fp8_e4m3", JBLAS_DTYPE::F8_E4M3},
+    {"fp8_e5m2", JBLAS_DTYPE::F8_E5M2}};
+static std::map<std::string, JBLAS_DTYPE> scale2jblasdt_map{{"fp32", JBLAS_DTYPE::F32},
+                                                            {"fp8_e8m0", JBLAS_DTYPE::F8_E8M0}};
+
 void dispatch_woq_task(woq_config_param* p, woq_runtime_ctx* ctx, WOQ_TASK task);
 void jblas_packq(woq_packq_param* p, woq_packq_ctx* ctx);
 void set_woq_workspace(torch::Tensor* workspace);
