+++ intel_extension_for_transformers/llm/operator/csrc/dispatcher/src/jblas_weightonly_dispatcher.cpp
@@ -34,13 +34,6 @@ inline void set_nk(woq_runtime_ctx* ctx, torch::Tensor* tensor) {
   ctx->k = ctx->transpose ? static_cast<int>(tensor->sizes()[1]) : static_cast<int>(tensor->sizes()[0]);
 }
 
-static std::map<std::string, JBLAS_DTYPE> wei2jblasdt_map{
-    {"int4_clip", JBLAS_DTYPE::S4_CLIP}, {"int4_fullrange", JBLAS_DTYPE::S4_FULLRANGE},
-    {"nf4", JBLAS_DTYPE::F4_NF4},        {"fp4_e2m1_bnb", JBLAS_DTYPE::F4_BNB},
-    {"fp4_e2m1", JBLAS_DTYPE::F4_E2M1},  {"fp8_e4m3", JBLAS_DTYPE::F8_E4M3},
-    {"fp8_e5m2", JBLAS_DTYPE::F8_E5M2}};
-static std::map<std::string, JBLAS_DTYPE> scale2jblasdt_map{{"fp32", JBLAS_DTYPE::F32},
-                                                            {"fp8_e8m0", JBLAS_DTYPE::F8_E8M0}};
 static void* woq_workspace = nullptr;
 static int64_t workspace_size = 0;
 
@@ -400,59 +393,6 @@ void dispatch_woq_task(woq_config_param* p, woq_runtime_ctx* ctx, WOQ_TASK task)
 void set_woq_workspace(torch::Tensor* workspace) {
   woq_workspace = workspace->data_ptr();
   workspace_size = workspace->element_size() * workspace->numel();
-}
-
-template <class GemmCore, JBLAS_ISA ISA>
-void execute_qpack(woq_packq_param* p, woq_packq_ctx* ctx) {
-  using proB = jblas::prologue_b::gemm::WeightKBlockNInteger<GemmCore, ISA>;
-  static proB ker;
-  bool asym = p->alg == "asym";
-  auto qpackw = ker.createStorage(ctx->n, ctx->k, p->group_size, wei2jblasdt_map[p->weight_type],
-                                  scale2jblasdt_map[p->scale_type], JBLAS_DTYPE::BF16, asym);
-  if (p->enable_act_shuffle) ker.enableShuffle(&qpackw);
-  *(ctx->output) = torch::empty(qpackw.mSize, torch::kInt8);
-  qpackw.assign(ctx->output->data_ptr<int8_t>());
-  if (p->enable_act_shuffle)
-    ker.setShuffleIndices(ctx->g_idx->data_ptr<int>(), &qpackw, &dispatcher_utils::DefaultThreading);
-  ker.packQWeight(ctx->n, ctx->k, ctx->qweight->data_ptr<int8_t>(), ctx->n, ctx->scale->data_ptr<float>(),
-                  asym ? ctx->zp->data_ptr<int8_t>() : nullptr, &qpackw, &dispatcher_utils::DefaultThreading);
-}
-
-void jblas_packq(woq_packq_param* p, woq_packq_ctx* ctx) {
-  TORCH_CHECK(p->weight_type == "int8" || p->weight_type == "int4_clip" || p->weight_type == "int4_fullrange",
-              "Qbits: only support Integer WOQ in PACKQ");
-
-  if (p->compute_type == "int8") {
-    if (dispatcher_utils::check_amx() && p->group_size % jblas::gemm::ICoreRowNAmxint8KBlock<48, 16>::KTILE == 0) {
-      return execute_qpack<jblas::gemm::ICoreRowNAmxint8KBlock<48, 16>, JblasAMX_INT8>(p, ctx);
-    }
-    if (dispatcher_utils::check_avx512_vnni() &&
-        p->group_size % jblas::gemm::ICoreRowNAvx512vnniKBlock<48, 4>::KTILE == 0) {
-      return execute_qpack<jblas::gemm::ICoreRowNAvx512vnniKBlock<48, 4>, JblasAVX512_VNNI>(p, ctx);
-    }
-    if (dispatcher_utils::check_avx_vnni() && p->group_size % jblas::gemm::ICoreRowNAvxvnniKBlock<48, 2>::KTILE == 0) {
-      return execute_qpack<jblas::gemm::ICoreRowNAvxvnniKBlock<48, 2>, JblasAVX_VNNI>(p, ctx);
-    }
-    TORCH_CHECK(false, "Qbits: Illegal config in int8 compute_type, blocksize:", p->group_size,
-                ", ISA support vnni:", dispatcher_utils::check_avx_vnni());
-  }
-  if (p->compute_type == "fp32") {
-    if (dispatcher_utils::check_avx512f()) {
-      return execute_qpack<jblas::gemm::SCoreRowNAvx512f<48, 8>, JblasAVX512F>(p, ctx);
-    }
-    if (dispatcher_utils::check_avx2()) {
-      return execute_qpack<jblas::gemm::SCoreRowNAvx2<48, 2>, JblasAVX2>(p, ctx);
-    }
-    TORCH_CHECK(false, "Qbits: device ISA must support AVX2 when compute_type==fp32");
-  }
-  if (p->compute_type == "bf16") {
-    if (dispatcher_utils::check_amx()) {
-      return execute_qpack<jblas::gemm::HCoreRowNAmxbf16<64, 16>, JblasAMX_BF16>(p, ctx);
-    }
-    TORCH_CHECK(false, "Qbits: device ISA must support AMX-BF16 when compute_type==bf16");
-  }
-  TORCH_CHECK(false, "Qbits: unsupported jblas_config, compute_type:", p->compute_type,
-              ", weight_type:", p->weight_type + ", blocksize:", p->group_size);
 }
 
 }  // namespace woq
